+++ packages/expo-camera/ios/Current/CameraView.swift
@@ -179,7 +179,7 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
       self.session.beginConfiguration()
       let preset = self.pictureSize.toCapturePreset()
       if self.session.canSetSessionPreset(preset) {
-        self.session.sessionPreset = preset
+        self.session.sessionPreset = self.mode == .video ? preset : .photo
       }
       self.session.commitConfiguration()
     }
@@ -225,6 +225,7 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
         if self.videoFileOutput == nil {
           self.setupMovieFileCapture()
         }
+        self.updateSessionAudioIsMuted()
       } else {
         self.cleanupMovieFileCapture()
       }
@@ -252,6 +253,7 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
         self.photoOutput = photoOutput
       }
 
+      self.session.sessionPreset = self.mode == .video ? self.pictureSize.toCapturePreset() : .photo
       self.addErrorNotification()
       self.changePreviewOrientation()
     }
@@ -652,6 +654,7 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
       if session.outputs.contains(videoFileOutput) {
         self.session.beginConfiguration()
         session.removeOutput(videoFileOutput)
+        self.session.sessionPreset = .photo
         self.videoFileOutput = nil
         self.session.commitConfiguration()
       }
@@ -692,10 +695,6 @@ public class CameraView: ExpoView, EXCameraInterface, EXAppLifecycleListener,
 
     videoRecordedPromise = nil
     videoCodecType = nil
-
-    if session.sessionPreset != pictureSize.toCapturePreset() {
-      updateSessionPreset(preset: pictureSize.toCapturePreset())
-    }
   }
 
   func setPresetCamera(presetCamera: AVCaptureDevice.Position) {